pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.6",
    "boto3>=1.42.34",
    "boto3-stubs[dynamodb]>=1.42.34",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "mangum>=0.20.0",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "starlette>=0.50.0",
    "uuid6>=2025.0.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
